remove ctor init checking
I realized that is kinda dumb.

So, the only way a different thread can access this variable, declared
in some function, is for that thread to be created after the
initialization of the object inside the thread.

Which means, before the initialization of a new thread, the constructor
has already finished its work. Because, inside the same thread, shits
still happen sequentially.

That means, all those checkings are, dumb.

That being said, I should add a note:

If a new thread is to use the same allocator, it must make its own copy
of the allocator (preferably using the copy constructor). The copy still
points to the same memory block, so it's rather cheap to construct a
copy.

commit-by: Huy Nguyen